Four clippy lints in the new creature examples
CI gates `cargo clippy --all-targets -- -D warnings`, and the creature merge
brought in three example files that had not been through it: a redundant
`i as i32` where `i` is already `i32`, a range loop indexing the array it is
iterating, a collapsible `if`, and a `.map()` on an `Option` used for its
side effect. All mechanical, none of them change behaviour, and the examples
still build.

Found while verifying the merge of that work with mine rather than by
looking for it -- flagging because it means clippy was not run against that
branch, and it is the CI job that gates rather than informs.

+++ examples/creature_probe.rs
@@ -68,7 +68,7 @@ fn main() {
         }
     }
     for i in 0..55 {
-        world.plant_ant(20 + i as i32 * 2, floor - 1);
+        world.plant_ant(20 + i * 2, floor - 1);
     }
 
     println!("creature probe: {frames} frames, reporting {ants} ants every {every}");
